+++ src/main/java/com/kamilkazmierczak/Interfaces/IReceipt.java
@@ -8,7 +8,7 @@ import java.util.Map;
  * Created by Kamil on 07.05.2017.
  */
 public interface IReceipt {
-    long getTotalSum() throws ReceiptNotClosedException;
+    double getTotalSum() throws ReceiptNotClosedException;
     void addProduct(IProduct product);
     Map<IProduct, Integer> getAllProducts();
     void closeReceipt();
